feat: :construction: Countdown finished, added controls in modview
Finished the "Stream-start"-countdown, added controls in the Modview-panel to allow starting/pausing and changing the duration of the countdown. - Fixed some styling, that allowed the modview to accurately show the previews in miniature format, while keeping the same size in browser/OBS + Made the background video fill out the page, so there's no gaps. Overlay can now be accurately used on 1440P too. - Added alias-paths to tsconfig, vite.config, added @typed/node for this, removed a lot of relative-pathing.